Refactor keymap type from uint16_t to platform_keycode_t in test files
- Updated test cases in `test_oneshot.cpp`, `test_scenario.hpp`, `test_tap_dance_basic_state_machine.cpp`, `test_tap_dance_comprehensive.cpp`, `test_tap_dance_interrupt_flavors.cpp`, `test_tap_dance_nested_layers.cpp`, and `test_tap_dance_roll_over_previous_press.cpp` to use `platform_keycode_t` instead of `uint16_t` for keycode definitions and keymap structures.
- This change ensures consistency across the codebase and aligns with the new keycode type definition.

+++ src/pipeline_oneshot_modifier.c
@@ -32,7 +32,7 @@ void pipeline_oneshot_modifier_callback_process_data(pipeline_virtual_callback_p
     }
     if (global_status->modifiers != 0 && key_event->keycode <= 0xFF && key_event->is_press == true) {
         actions->mark_as_processed_fn();
-        uint16_t modifiers = global_status->modifiers;
+        uint8_t modifiers = global_status->modifiers;
         if (modifiers & MACRO_KEY_MODIFIER_LEFT_SHIFT) {
             actions->report_press_fn(PLATFORM_KC_LEFT_SHIFT);
         }
